refactor: move secret related components to secret folder
